bash: Remove trailing ';' from PROMPT_COMMAND.
Previously we ensured that we always had a trailing ';' after every
command in PROMPT_COMMAND. We now only add a ';' after the previous
command (if the variable was previously set).

We also explicitly unset the PROMPT_COMMAND that we get from the OS,
previously that was done implicitly, by not including it in the first
place we set it (even though that command was conditionally executed).